Move some files.
