+++ src/main/java/net/lewmc/essence/Essence.java
@@ -10,6 +10,14 @@ import net.lewmc.essence.commands.EssenceCommands;
 import net.lewmc.essence.commands.GamemodeCommands;
 import net.lewmc.essence.commands.stats.*;
 import net.lewmc.essence.commands.teleportation.*;
+import net.lewmc.essence.commands.teleportation.home.DelhomeCommand;
+import net.lewmc.essence.commands.teleportation.home.HomeCommand;
+import net.lewmc.essence.commands.teleportation.home.HomesCommand;
+import net.lewmc.essence.commands.teleportation.home.SethomeCommand;
+import net.lewmc.essence.commands.teleportation.warp.DelwarpCommand;
+import net.lewmc.essence.commands.teleportation.warp.SetwarpCommand;
+import net.lewmc.essence.commands.teleportation.warp.WarpCommand;
+import net.lewmc.essence.commands.teleportation.warp.WarpsCommand;
 import net.lewmc.essence.events.DeathEvent;
 import net.lewmc.essence.events.JoinEvent;
 import net.lewmc.essence.events.PlayerDamageEvent;
