+++ src/main/java/net/lewmc/essence/commands/teleportation/warp/DelwarpCommand.java
@@ -1,4 +1,4 @@
-package net.lewmc.essence.commands.teleportation;
+package net.lewmc.essence.commands.teleportation.warp;
 
 import net.lewmc.essence.utils.FileUtil;
 import net.lewmc.essence.utils.LogUtil;
@@ -8,7 +8,6 @@ import net.lewmc.essence.utils.PermissionHandler;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
-import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
